+++ CHANGELOG.md
@@ -12,6 +12,7 @@ v0.15.19 (??.??.????)
  * new Wrapper function tx_rnbase_parameters::getPostOrGetParameter for _GP
  * new Wrapper function tx_rnbase_util_Link::linkThisScript
  * new Wrapper function tx_rnbase_util_Files::writeFile
+ * new Wrapper function tx_rnbase_util_Strings::removeXSS
 
 v0.15.18 (02.12.2015)
  * Bugfix: tx_rnbase_util_Dates::date_mysql2tstamp liefert nun immer einen Integer und nicht NULL im Fehlerfall